fixed all the errors that appeared before. because it was using the vector3d struct instead of the object

+++ vector3d.h
@@ -5,18 +5,18 @@
 class Vector3D
 {
   public:
-    float x, y, z;
+    double x, y, z;
     //default constructor
-    Vector3D(float X = 0, float Y = 0, float Z = 0){ x = X; y = Y; z = Z;}
+    Vector3D(double X = 0, double Y = 0, double Z = 0){ x = X; y = Y; z = Z;}
     ~Vector3D(){};
     //calculate and return the magnitude of this vector
-    float GetMagnitude()
+    double GetMagnitude()
     { return sqrtf(x * x + y * y + z * z); }
     //multiply this vector by a scalar
-    Vector3D operator*(float num) const
+    Vector3D operator*(double num) const
     { return Vector3D(x * num, y * num, z * num); }
     //pass in a vector, pass in a scalar, return the product
-    friend Vector3D operator*(float num, Vector3D const &vec)
+    friend Vector3D operator*(double num, Vector3D const &vec)
     { return Vector3D(vec.x * num, vec.y * num, vec.z * num); }
     //add two vectors
     Vector3D operator+(const Vector3D &vec) const
@@ -26,14 +26,14 @@ class Vector3D
     { return Vector3D(x - vec.x, y - vec.y, z - vec.z); }
     //normalize this vector
     void normalizeVector3D(){
-      float magnitude = sqrtf(x * x + y * y + z * z);
+      double magnitude = sqrtf(x * x + y * y + z * z);
       x /= magnitude;
       y /= magnitude;
       z /= magnitude;
     }
 
     //calculate and return dot product
-    float dotVector3D(const Vector3D &vec) const
+    double dotVector3D(const Vector3D &vec) const
     { return x * vec.x + y * vec.y + z * vec.z; }
 
     //calculate and return cross product
